Alternative way of snapping bounds to the grid
Changes to take into account changes in grid size, when snapping positions to grid.
Specifically when trying to snap from a smaller to a larger grid.

Pressing and holding the Y key, by default, will activate this behaviour.
Alternatively there's a toggle in the settings that can permanently turn this on
so that the activation button doesn't need to be pressed.
This is off by default.