Fix build error when build inside prefab instance
On Unity 2018.3, the new nested prefab feature came.
This prohibits removing gameobjects in a prefab instance.
So, we should remove old meshes under the mesh group not in a prefab instance but in an original prefab.

Besides, we may create new gameobjects in the original prefab, too.
But creating them in the prefab instance isn't prohibited, so I leave this unchanged.
Because of this, you should apply prefab changes manually after rebuild.